fix(libs/header/sync): resolve head synchronization issue (#1878)
Because syncedHead and writeHead inside of Store are not synchronized, we can have a case we set a sync head as an old header. This happens when we receive an old header from the network.

This is only fixing PR that has some ugliness on it. The refactoring PR is to follow.