fix(header/sync): add drift to recency detection (#1922)
Pretty often, when you start a node for the first time, there is a warning complaining about an unsynced trusted peer, while we know that the peer is 100% synced. This warning also means we make an additional Head request for the wrong reason. The fix is introducing a half-block time drift. I've tested it like 20 times and couldn't reproduce the warning. We also don't need to be super precise about head to worry about this additional half. The main goal is to get the head quickly enough after being out of sync for a day or more.