fix(store): make storePrefix actually optional (#69)
When we set the default value on the private config field while requiring it, we brake the serialization of the Parameters struct.
Particularly when we deserialize Parameters, the storePrefix is empty while the validation requires it. Because it's private, there is also no way to set it on the node, so this fix is necessary. This is also good learning that we should be cautious while introducing such optinonals