fix(store): greedy heightSub cancellation
So I was playing around with the similar code and taking heightSub as an inspiration. While writing that code I realized that heightSub has a nasty bug.
Basically a subscription cancellation removes the others active subs for the same height leaving them blocked forever.
We could observe some abnormal blockings in the past and this may explain it.