fix(store): properly update store head (#207)
## Overview

The main idea of this PR is to make `Store[H].Head` working properly, to
be precise: returning head that was written to the disk (*). Along with
that `heightSub.height` is increased monotonically to prevent bugs when
we have store appends out of order.

To test everything I'm adding 2 new tests: one that verifies out of
order appends and another when which does this concurrently. Which
helped to find 2 or even 3 edge cases during coding.

Fixes #201