fix(p2p): catch rare nil Header (#234)
A subscription might be started without the PubSub validator fully
registered, causing nil pointer deref. This was observed as a flake in
tests and sometimes even on the celestia node start. It's time to fix
this issue altogether.

The naive fix would check for nil, discarding a valid header message. On
the other hand, this fix ensures proper order of events, guaranteeing
that a valid message is never processed by subscription before the
user's header verifier is set.

Related to https://github.com/celestiaorg/celestia-node/issues/4001