feat(sync): non-recursive singleflight protection for Head (#229)
Closes: https://github.com/celestiaorg/go-header/issues/154

This PR adds singleflight protection to prevent duplicate header
requests in the Exchange component. The focus is on minimizing redundant
requests, especially around the Head method, which is most likely to
experience simultaneous calls.

I moved out single-flight protection out of exchange to be in sync.Syncer. This way the single-flight protection works for every Head implementation: p2p, core. Besides, I avoid using the x/sync/singleflight as it does not handle context cancellation.

This also fixes an extremely rare but possible issue with subjective initialization similar to #247 (cc @walldiss). This issue is caused by to be removed syncGetter when used together with recursion. The fix is to move singleflight handling out of sync.Syncer, it should not be responsible for this. This way we fix the issue, without the need for read/write refactoring in sync_head.go we discussed.

---------

Co-authored-by: Wondertan <hlibwondertan@gmail.com>
Co-authored-by: IliaBulavintsev <ibulavincev@finch-melrose.com>