fix(sync): discard headers failed bifurcation (#269)
#219 introduced bifurcation however we missed there an issue with
setting a sync target as discovered in
https://github.com/celestiaorg/go-header/pull/268#issuecomment-2774042939.
Basically, a soft failed header failing bifurcation would still be used
a sync target. This is wrong and defeats the purpose of bifurcation.

This patch changes this so that bifurcation is used as source of truth
for headers we are unsure about(soft failed). If such header appears, we
bifurcate and get to the bottom whether its valid or not, s.t. the
header never enters syncing pipeline in pending cache and doesn't have
to be treated differently and thus simplify syncing flow.

Besides, this change sets all the discovered which were headers along
the bifurcation as subjective head, facilitating the syncing progress.