fix(store): support complete erase (#285)
Previously Store allowed deleting up to the Head. This change allows
deleting including the head too, ensuring proper cleanup of Store's
state in such a case. If head is `5` and `DeleteTo(6)` is called, the
store will handle this case now.

Additionally, if head was `5` and `DeleteTo(7)` is called, while `7`and
`8` exists in the store, they will become Tail and Head respectively.

Closes #284