feat(sync): maintain tail (#265)
Adds `Tail` method to `Syncer` that can initialize a node with a height
estimated header or using the configured height(or hash) as the new
`Tail`. Besides, it can gracefully handle reconfiguration of height/hash
upon restarts.

~This change deliberately leaves some tech debt behind. The new Tail
method has quite a lot of branchiness and could be teared
apart/rewritten reducing nesting and improving readability. Some more
specific debts are marked with a TODO. All of this is not "clean" and
simply achieves the goal. The future backward sync will anyway require
us to refactor this logic, so there is minimum value in cleaning things
now.~

Nvm, I end up refactoring it for the sake of reviewers sanity and code
maintainability. There are still TODOs left for things that are
literally blocked on bsync or else, but everything else was uniformly
integrated into the existing code.

Depends on #283 and #285

Closes #258 and #259