fix(sync): avoid Tail double requesting (#290)
To avoid requesting tail twice, I had to extract new `networkHead`
method out of `Head` method. This refactoring lead to a few adjacent
refactors:
* avoids double verification of a recent head request. Previously, we
verified it in exchange and then in the syncer's code. This double
verification is actually a bad one that verifies all the signatures and
not just `NextValidatorHash`
* prevents duplicate request right after subjective init for a more
recent head, if the subjective initialization yielded non recent header.
* extract `localHead` out of `subjectiveHead`
* improves logging