feat(store): allow partial init (#327)
Previously, if either head or tail is not found for whatever reason, the
store will report being empty, causing the node to reinitialize and sync
from scratch. This change enables it to recover either the head or the
tail if the other is missing. For example, if the tail is missing due to
an interrupted delete, the head will be retrieved successfully upon
start. The tail will then be recovered via the receding tail
functionality. Although if the chain is too long, receding may take too
long making it hard for the node to stop, thus #326

This was tested manually, but it still needs a unit test. It will happen
in a separate PR