fix(sync): subj init death loop (#324)
During subjective init, we verify the head against the tail. However,
the tail may be configured to be outside of the trusting window, causing
subjective init "deathloop".

Once the tail is requested, it is immediately stored as the first-ever
header. However, this tail is also a head for a brief moment before the
network head gets applied in `incomingNewHead`. And when `verify` inside
of it calls `subjectiveHead`, the temporary head tail is considered
expired, causing re-requesting of the new head. However, this new head
fails verification as it's the same as the network head awaiting to be
applied. Then this whole process goes nuts in cycles.

This is actually also a solution for the original problem described in
#317, but it does not solve the secondary problem in the EDIT section,
so we still need to keep #317