perf(store): batch reads (#331)
This PR shaves another 30s when we prune the whole chain. It achieves
this by applying the same technique as in #328, but for reads.
Previously, the underlying datastore would create a read-only
transaction for every singular read. As shown in pyroscope profiles,
this puts pressure on badger as creating new transactions ain't cheap
and involves complex synchronization to ensure atomicity between
parallel transactions.

With this insight in mind, I also integrated read batching in other
places where this must be beneficial: advancing, receding, range, and
singular header reads.

More noticeable changes:
* Fixes a silent bug that wouldn't delete samples on the Pruner side
* Fixes a bug in the upstream contextds change
* Adds tracking for missing headers that were not found during deletion